scanner/requirements.txt
yfinance>=0.2.40
pandas>=2.2.0
gspread>=6.0.0
google-auth>=2.25.0
resend>=4.0.0
requests>=2.31.0
